% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncor_ind.R
\name{ncor_ind}
\alias{ncor_ind}
\title{A Negative Control Outcome Regression for
Eliminating Unobserved Confounding in Time-series Studies}
\usage{
ncor_ind(
  data = data,
  pre_outc_name = "y1",
  expo_namem = "x",
  post_outc_name = "y3",
  centre = "centre",
  method = c("regression", "IVW"),
  boot_no = NULL
)
}
\arguments{
\item{data}{an optional data frame containing the variables in the model.}

\item{pre_outc_name}{the name of pre-exposure outcome}

\item{expo_namem}{the name of exposure}

\item{post_outc_name}{the name of post-exposure outcome}

\item{centre}{the number of reacher centre or time series fragments}

\item{method}{method of estimation}

\item{boots_no}{the number of bootstrap for IVW estimation}
}
\value{
\code{causal} the casual effect estimation

\code{lag} the lag causal effect estimation
}
\description{
A Negative Control Outcome Regression for
Eliminating Unobserved Confounding in Time-series Studies
}
\examples{

NN <- 10
data_total <- NULL
for(i in 1:NN){
c0 <- 0.5
dia <- 0.5
diff_c1 <- 0
r <- 0.5
c1 <- rnorm(1,0.5,1)
c2 <- rnorm(1,0.5,1)
N <- 100
Sigma <- matrix(c(1,r,r^2,r,1,r,r^2,r,1),3,3)
u <- MASS::mvrnorm(n=N, mu=rep(0,3), Sigma=Sigma)
colnames(u) <- c('u1','u2','u3')
u <- data.frame(u)
u1 <- u$u1
u2 <- u$u2
u3 <- u$u3

eps_y1 <- rnorm(N,0,0.1)
eps_y3 <- rnorm(N,0,0.1)
eps_x2 <-  rnorm(N,0,0.1)
y1 <- c1*u1 + eps_y1
x2 <- c2*u2 + eps_x2
y3 <- c0*x2 + c1*u3+diff_c1*u3 + dia*y1 + eps_y3
data_cenre <- data.frame(x2,u1,u2,u3,y1,y3,i)
data_total <- rbind(data_total,data_cenre)

}

model <- ncor_ind (data=data_total, pre_outc_name='y1',expo_namem = 'x2',
                  post_outc_name='y3',centre='i',method='IVW',boot_no=1000)

model


}
