% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncor_summary.R
\name{ncor_summary}
\alias{ncor_summary}
\title{A Negative Control Outcome Regression for
Eliminating Unobserved Confounding in Time-series Studies}
\usage{
ncor_summary(
  data = data,
  coef_y3_x = "coef1",
  coef_y1_x = "coef2",
  se = "se",
  boot_no = 1000
)
}
\arguments{
\item{data}{an optional data frame containing the variables in the model.}

\item{coef_y3_x}{the coef of exposure on post-exposure outcome}

\item{coef_y1_x}{the coef of exposure on pre-exposure outcome}

\item{se}{the standard error of coef_y3_x}

\item{boots_no}{the number of bootstrap for estimation}
}
\value{
\code{causal} the casual effect estimation

\code{lag} the lag causal effect estimation
}
\description{
A Negative Control Outcome Regression for
Eliminating Unobserved Confounding in Time-series Studies
}
