% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCor.R
\name{NCor}
\alias{NCor}
\title{Wermuth's Correlation}
\usage{
NCor(X, Y = NULL, alpha = 0.1, digits = 5, nominal = "rc", method = "gamma")
}
\arguments{
\item{X}{an n x 1 character vector or a contingency table with at least one nominal random variable.}

\item{Y}{NULL (default) or a n x 1 character or numeric vector.}

\item{alpha}{confidence level for the returned confidence interval. FALSE yields the coefficient without confidence intervals.}

\item{digits}{only applicable if a contingency table of probabilities is supplied. Each probability gets multiplied by 10^digits to obtain an artificial sample/population size.}

\item{nominal}{determines whether the rows ("r"), the columns ("c") or the rows and the columns ("rc") of the resulting contingency table should be treated as nominal. If X and Y are supplied separately, X refers to the rows and Y to the columns.}

\item{method}{determines the coefficient to be used. Default is Goodman-Kruskal's gamma \insertCite{Goodman1954}{NCor}.}
}
\value{
The value of the coefficient.
}
\description{
\code{NCor()} computes proper correlation coefficients for nominal random variables together with optional confidence intervals and independence tests.
}
\examples{
tab <- matrix(c(38, 0, 75, 16, 0, 43, 86, 84, 60), ncol = 3)
colnames(tab) <- c("Afghanistan", "Somalia", "Iran")
rownames(tab) <- c("Islam", "Judaism", "Christianity")
NCor(tab)
}
\references{
\insertRef{Goodman1954}{NCor}
}
